requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
camoufox==0.4.11
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
slowapi==0.1.9
httpx==0.25.0